refactor: move downloadSong and repairSongs into src/core/download
